refactor(types): import contract types directly from swissknife-types
Drop the pass-through dtos re-export files; handlers, use cases and services
now import requests/responses straight from the crate. Move the three edge
responses with no entity (SignInResponse, NostrNIP05Response, ErrorResponse)
into a response module, and collapse LnUrlCallbackResponse into the
LnUrlCallback entity (typed success_action, camelCase wire shape).

GET /v1/me/lightning-address returns Option<LnAddress> instead of the
WalletLnAddressResponse wrapper. AppConfig stays in dtos (not a wire type).
