+++ crates/swissknife-types/src/lnurl.rs
@@ -1,11 +1,19 @@
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
-#[derive(Debug)]
+/// LNURL-pay callback response. Carries the invoice to pay and how to behave on
+/// success. Wire shape follows LUD-06 (camelCase fields).
+#[derive(Debug, Clone, Serialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
 pub struct LnUrlCallback {
+    /// bech32-serialized Lightning invoice
+    #[schema(example = "lnbcrt1m1png24kasp5...")]
     pub pr: String,
+    /// An optional action to be executed after successfully paying an invoice
     pub success_action: Option<LnUrlSuccessAction>,
+    /// An optional flag to let a wallet know whether to persist the link from step 1, if null should be interpreted as true
     pub disposable: Option<bool>,
+    /// array with payment routes, should be left empty if no routes are to be provided
     pub routes: Vec<String>,
 }
