--- src/application/dtos/api_key.rs
@@ -1,1 +1,0 @@
-pub use swissknife_types::CreateApiKeyRequest;
