--- src/application/dtos/auth.rs
@@ -1,12 +1,0 @@
-use serde::Serialize;
-use utoipa::ToSchema;
-
-pub use swissknife_types::{SignInRequest, SignUpRequest};
-
-/// Sign In Response
-#[derive(Debug, Serialize, ToSchema)]
-pub struct SignInResponse {
-    /// JWT token
-    #[schema(example = "eyJ0eXAiOiJKV1QiLCJhbGciOiJ...")]
-    pub token: String,
-}
