--- src/application/dtos/bitcoin.rs
@@ -1,1 +1,0 @@
-pub use swissknife_types::NewBtcAddressRequest;
