--- src/application/dtos/invoices.rs
@@ -1,1 +1,0 @@
-pub use swissknife_types::NewInvoiceRequest;
