+++ src/application/dtos/mod.rs
@@ -1,23 +1,3 @@
-mod api_key;
-mod auth;
-mod bitcoin;
 mod config;
-mod error;
-mod invoices;
-mod ln_address;
-mod lnurl;
-mod nostr;
-mod payments;
-mod wallet;
 
-pub use api_key::*;
-pub use auth::*;
-pub use bitcoin::*;
 pub use config::*;
-pub use error::*;
-pub use invoices::*;
-pub use ln_address::*;
-pub use lnurl::*;
-pub use nostr::*;
-pub use payments::*;
-pub use wallet::*;
