--- src/application/dtos/payments.rs
@@ -1,1 +1,0 @@
-pub use swissknife_types::SendPaymentRequest;
