--- src/application/dtos/wallet.rs
@@ -1,1 +1,0 @@
-pub use swissknife_types::RegisterWalletRequest;
