+++ src/domains/invoice/invoice_handler.rs
@@ -8,13 +8,14 @@ use axum::{
 use utoipa::OpenApi;
 use uuid::Uuid;
 
+use swissknife_types::{ErrorResponse, NewInvoiceRequest};
+
 use crate::{
     application::{
         docs::{
             BAD_REQUEST_EXAMPLE, FORBIDDEN_EXAMPLE, INTERNAL_EXAMPLE, NOT_FOUND_EXAMPLE, UNAUTHORIZED_EXAMPLE,
             UNPROCESSABLE_EXAMPLE,
         },
-        dtos::{ErrorResponse, NewInvoiceRequest},
         entities::AppServices,
         errors::ApplicationError,
     },
