+++ src/domains/wallet/wallet_handler.rs
@@ -9,13 +9,14 @@ use axum_extra::extract::Query;
 use utoipa::OpenApi;
 use uuid::Uuid;
 
+use swissknife_types::{ErrorResponse, RegisterWalletRequest};
+
 use crate::{
     application::{
         docs::{
             BAD_REQUEST_EXAMPLE, FORBIDDEN_EXAMPLE, INTERNAL_EXAMPLE, NOT_FOUND_EXAMPLE, UNAUTHORIZED_EXAMPLE,
             UNPROCESSABLE_EXAMPLE,
         },
-        dtos::{ErrorResponse, RegisterWalletRequest},
         entities::AppServices,
         errors::ApplicationError,
     },
